// libs/features/collection/billing/src/lib/shared/models/itemized-contribution-month-wrapper.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms .
 */
import { ItemizedContributionMonth } from './itemized-contribution-month';

export class ItemizedContributionMonthWrapper {
  thirdPartyItemizedBillBreakDown: ItemizedContributionMonth[] = [];
  availableCount: number = undefined;
}